# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Design Print Production Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
    post:
      operationId: createQuotation
      tags: [quotations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
    patch:
      operationId: updateQuotation
      tags: [quotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotations/{id}/convert:
    post:
      operationId: convertQuotationToOrder
      tags: [quotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Sales order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"

  # Sales Orders
...
  /machines:
    get:
      operationId: listMachines
      tags: [machines]
      responses:
        "200":
          description: List of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
    post:
      operationId: createMachine
      tags: [machines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
  /machines/{id}:
    get:
      operationId: getMachine
      tags: [machines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Machine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
    patch:
      operationId: updateMachine
      tags: [machines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
    delete:
      operationId: deleteMachine
      tags: [machines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Materials / Inventory
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    patch:
      operationId: updateMaterial
      tags: [materials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Job Costing
...
  /job-orders/{id}/qc:
    get:
      operationId: getQCChecklist
      tags: [qc]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QC checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QCChecklist"
    put:
      operationId: upsertQCChecklist
      tags: [qc]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QCChecklistInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QCChecklist"

  # Dashboard
...
components:
  schemas:
...
    Quotation:
      type: object
      required: [id, quotationNumber, customerId, productType, quantity, unitPrice, totalPrice, status, createdAt]
      properties:
        id:
          type: integer
        quotationNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        productType:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        paymentTerms:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    QuotationInput:
      type: object
      required: [customerId, productType, quantity, unitPrice]
      properties:
        customerId:
          type: integer
        productType:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        paymentTerms:
          type: string
        deliveryDate:
          type: string
        status:
          type: string
        notes:
          type: string
    QuotationUpdate:
      type: object
      properties:
        productType:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        paymentTerms:
          type: string
        deliveryDate:
          type: string
        status:
          type: string
        notes:
          type: string
    SalesOrder:
      type: object
      required: [id, orderNumber, customerId, productType, quantity, orderStatus, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        quotationId:
          type: ["integer", "null"]
        quotationNumber:
          type: ["string", "null"]
        productType:
          type: string
        quantity:
          type: integer
        deliveryDate:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [unpaid, partial, paid]
        orderStatus:
          type: string
          enum: [pending, confirmed, in_production, completed, delivered, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Machine:
      type: object
      required: [id, name, productionLine, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        productionLine:
          type: string
        status:
          type: string
          enum: [available, running, maintenance, stopped]
        currentJobId:
          type: ["integer", "null"]
        currentJobNumber:
          type: ["string", "null"]
        operatorId:
          type: ["integer", "null"]
        operatorName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    MachineInput:
      type: object
      required: [name, productionLine]
      properties:
        name:
          type: string
        productionLine:
          type: string
        status:
          type: string
        operatorId:
          type: integer
        notes:
          type: string
    MachineUpdate:
      type: object
      properties:
        name:
          type: string
        productionLine:
          type: string
        status:
          type: string
        currentJobId:
          type: integer
        operatorId:
          type: integer
        notes:
          type: string
    Material:
      type: object
      required: [id, name, category, unit, currentQuantity, minimumQuantity, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentQuantity:
          type: number
        minimumQuantity:
          type: number
        supplier:
          type: ["string", "null"]
        costPerUnit:
          type: ["number", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
    MaterialInput:
      type: object
      required: [name, category, unit, currentQuantity, minimumQuantity]
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentQuantity:
          type: number
        minimumQuantity:
          type: number
        supplier:
          type: string
        costPerUnit:
          type: number
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentQuantity:
          type: number
        minimumQuantity:
          type: number
        supplier:
          type: string
        costPerUnit:
          type: number
...
    QCChecklist:
      type: object
      required: [id, jobOrderId, result]
      properties:
        id:
          type: integer
        jobOrderId:
          type: integer
        colorAccuracy:
          type: ["boolean", "null"]
        sizeAccuracy:
          type: ["boolean", "null"]
        materialQuality:
          type: ["boolean", "null"]
        printingDefects:
          type: ["boolean", "null"]
        finishingQuality:
          type: ["boolean", "null"]
        packingQuality:
          type: ["boolean", "null"]
        finalApproval:
          type: ["boolean", "null"]
        result:
          type: string
          enum: [passed, rework, rejected, pending]
        notes:
          type: ["string", "null"]
        inspectedBy:
          type: ["string", "null"]
        inspectedAt:
          type: ["string", "null"]
    QCChecklistInput:
      type: object
      properties:
        colorAccuracy:
          type: boolean
        sizeAccuracy:
          type: boolean
        materialQuality:
          type: boolean
        printingDefects:
          type: boolean
        finishingQuality:
          type: boolean
        packingQuality:
          type: boolean
        finalApproval:
          type: boolean
        result:
          type: string
          enum: [passed, rework, rejected, pending]
        notes:
          type: string
...