# lib/api-zod/src/generated/types/salesOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Design Print Production Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { SalesOrderOrderStatus } from './salesOrderOrderStatus';
import type { SalesOrderPaymentStatus } from './salesOrderPaymentStatus';

export interface SalesOrder {
  id: number;
  orderNumber: string;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  quotationId?: number | null;
  /** @nullable */
  quotationNumber?: string | null;
  productType: string;
  quantity: number;
  /** @nullable */
  deliveryDate?: string | null;
  paymentStatus: SalesOrderPaymentStatus;
  orderStatus: SalesOrderOrderStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}